extract InnerParser.cs
